% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cid.R
\name{cid}
\alias{cid}
\alias{cid_dialog}
\alias{cid_csv}
\alias{cid_vector}
\alias{cid_clipboard}
\title{Search cattle information by ID}
\usage{
cid_dialog(encoding = getOption("encoding"))

cid_csv(input, output = "cattle_info.csv", append = T, col = 1,
  skip = 0, nrows = -1, ...)

cid_vector(input, output = "cattle_info.csv", append = T)

cid_clipboard(output = "cattle_info.csv", append = T)
}
\arguments{
\item{encoding}{Encoding of file names}

\item{input}{Path to a input csv file (\code{cid_csv}) or a vector (integer/numeric/character/factor) (\code{cid_vector}). A factor vector is interpretted 'literally'.}

\item{output}{Path to a output csv file.
If \code{NULL}, no csv file will be made.}

\item{append}{If an output file is already exist, append output file (T)
or create a new file (F). When \code{append = T}, ERROR rows in previous
output file will be removed.}

\item{col}{The column number of the input file contains cattle ID numbers.}

\item{skip, nrows, ...}{Be passed to \code{\link{read.csv}}}
}
\value{
A data.frame contains cattle information
}
\description{
Searches cattle information from https://www.id.nlbc.go.jp/ by cattle identification numbers and output a csv file.
}
\details{
Searches and downloads cattle information from database of National Livestock Breeding Center, Japan (https://www.id.nlbc.go.jp/) by cattle identification numbers and outputs a csv file contains cattle information.

\code{cid_dialog} opens dialog to choose an input file and an output file and set options. This one is most recommended.

\code{cid_csv} uses a csv file as input.

\code{cid_vector} uses a vector as input.

\code{cid_clipboard} uses the clipboard as input.

Search speed is about 10-12 cattles/min.
}
\examples{
\dontrun{
cid_dialog()
}

id <- c(1083079037, 0123456789, 0396006198, 1389725782)
cid_vector(input = id, output = NULL)
}
