% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ids.R
\name{load_ids}
\alias{load_ids}
\title{Loads cattle ID numbers from a csv file or clipboard}
\usage{
load_ids(input, use_clipboard = F, col = 1, skip = 0, nrows = -1, ...)
}
\arguments{
\item{input}{Path to a input csv file}

\item{use_clipboard}{If TRUE, clipboard is used instead of a input file}

\item{col}{The column number contains cattle ID numbers.
When use_clipboard = T, this parameter is ignored.}

\item{skip, nrows, ...}{Be passed to \code{\link{read.csv}}}
}
\value{
Loaderd cattle id numbers
}
\description{
Loads cattle ID numbers from a csv file or clipboard
}
