% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dialog.R
\name{make_dialog}
\alias{make_dialog}
\title{Makes dialog to choose a input file and a output path and starts scraping}
\usage{
make_dialog(encoding = getOption("encoding"))
}
\arguments{
\item{encoding}{Encoding of file names}
}
\description{
Makes dialog to choose a input file and a output path and starts scraping
}
