% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dialog.R
\name{ProgressBar}
\alias{ProgressBar}
\title{Progress bar}
\usage{
ProgressBar(gui_pb, ...)
}
\arguments{
\item{gui_pb}{Show progress bar in a window (T) or in console (F)}

\item{...}{Be passed to \code{\link{tcltk::tkProgressBar}} or
\code{\link{utils::txtProgressBar}}}
}
\description{
Progress bar
}
