% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nlbc.R
\name{save2csv}
\alias{save2csv}
\title{Save scraped information to a csv file}
\usage{
save2csv(info_50, output, env)
}
\arguments{
\item{info_50}{A table contains scraped information}

\item{output}{Path to the output file}

\item{env}{Envirionment which has a variable "info". (Don't set it manually.)}
}
\description{
Save scraped information to a csv file
}
