% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nlbc.R
\name{scrape_50}
\alias{scrape_50}
\title{Scrape 50 cattle}
\usage{
scrape_50(
  scrape_start,
  scrape_end,
  ids,
  output,
  lng_ids,
  fileEncoding,
  gui_pb,
  pb,
  env_nlbc
)
}
\arguments{
\item{scrape_start, scrape_end}{Define lines to scrape}

\item{ids}{A character vector of cattle ID number}

\item{output}{Path to the output file}

\item{lng_ids}{Length of ids}

\item{fileEncoding}{Encoding of the output file. See \code{\link{file}}.}

\item{gui_pb}{Show progress bar in a window (T) or in console (F)
(Don't set it manually)}

\item{pb}{Progress bar (Don't set it manually)}

\item{env_nlbc}{Parent frame (Don't set it manually)}
}
\description{
Scrape 50 cattle
}
