% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nlbc.R
\name{scrape_info}
\alias{scrape_info}
\title{Scrape information from a NLBC information page}
\usage{
scrape_info(page)
}
\arguments{
\item{page}{html page}
}
\description{
Scrape information from a NLBC information page
}
