% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nlbc.R
\name{scrape_info_cattle}
\alias{scrape_info_cattle}
\title{Scrape information of cattle from a NLBC information page}
\usage{
scrape_info_cattle(page)
}
\arguments{
\item{page}{html page}
}
\description{
Scrape information of cattle from a NLBC information page
}
