% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nlbc.R
\name{scrape_info_farm}
\alias{scrape_info_farm}
\title{Scrape information of farmer from a NLBC information page}
\usage{
scrape_info_farm(page)
}
\arguments{
\item{page}{html page}
}
\description{
Scrape information of farmer from a NLBC information page
}
