% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nlbc.R
\name{scrape_nlbc}
\alias{scrape_nlbc}
\title{Scrape NLBC websites}
\usage{
scrape_nlbc(
  ids,
  output = "cattle_info.csv",
  append = T,
  fileEncoding = getOption("encoding"),
  gui_pb = F
)
}
\arguments{
\item{ids}{Cattle ID numbers as a character vector}

\item{output}{Path to output file.
If \code{NULL}, no csv file will be made.}

\item{append}{If an output file is already exist, append output file (T)
or create a new file (F).}

\item{fileEncoding}{Encoding of the input/output file. See \code{\link{file}}.}

\item{gui_pb}{Show progress bar in a window (T) or in console (F)}
}
\description{
Scrape NLBC websites
}
