% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dialog.R
\name{setProgressBar}
\alias{setProgressBar}
\title{Set progress bar}
\usage{
setProgressBar(gui_pb, ...)
}
\arguments{
\item{gui_pb}{Show progress bar in a window (T) or in console (F)}

\item{...}{Be passed to \code{tcltk::setTkProgressBar} or
\code{utils::setTxtProgressBar}}
}
\description{
Set progress bar
}
