% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ids.R
\name{validate_ids}
\alias{validate_ids}
\title{Validates cattle ID numbers}
\usage{
validate_ids(ids)
}
\arguments{
\item{ids}{A vector contains cattle ID numbers.
If an factor vector is supplied, it's interpretted as literally.}
}
\value{
Cattle ID numbers as character
}
\description{
Validates cattle ID numbers
}
