% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nlbc.R
\name{write_log}
\alias{write_log}
\title{Output error-log}
\usage{
write_log(err_file, flag_nocattle, now_scraping, lng_ids)
}
\arguments{
\item{err_file}{Path to the error log file}

\item{flag_nocattle}{Flag whether an error had occured due to no cattle in the database.}

\item{lng_ids}{Length of IDs}

\item{now_scrapoing}{Index of ID of current scraping cattle}
}
\description{
Output error-log
}
