% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nlbc.R
\name{write_log}
\alias{write_log}
\title{Output error-log}
\usage{
write_log(
  err_file,
  flag_error,
  flag_nocattle,
  ids,
  now_scraping,
  lng_ids,
  err_catch
)
}
\arguments{
\item{err_file}{Path to the error log file}

\item{flag_error}{Flag whether an error had occurred}

\item{flag_nocattle}{Flag whether an error had occured due to no cattle in the database.}

\item{ids}{IDs to search}

\item{lng_ids}{Length of IDs}

\item{err_catch}{An object of class `try-error`}

\item{now_scrapoing}{Index of ID of current scraping cattle}
}
\description{
Output error-log
}
